services/emotion-inference/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
protobuf==4.25.3
pytest==7.4.3
